Fix 32-bit compilation, assuming local modifications to vulkan.h
* When the final header is ready we'll roll it into the source directly
  rather than assuming the SDK is installed anywhere, and apply a local
  change to get type-safety on 32-bit (which only works on C++). Similar
  to the GL header's modification for a typed GLenum.

+++ renderdoc/driver/d3d11/d3d11_context1_wrap.cpp
@@ -67,7 +67,7 @@ bool WrappedID3D11DeviceContext::Serialise_UpdateSubresource1(ID3D11Resource *pD
 			
 			if(WrappedID3D11Buffer::IsAlloc(DestResource))
 			{
-				srcLength = record->Length;
+				srcLength = (size_t)record->Length;
 
 				if(HasDestBox)
 					srcLength = RDCMIN((uint32_t)srcLength, pDstBox->right - pDstBox->left);
@@ -82,7 +82,7 @@ bool WrappedID3D11DeviceContext::Serialise_UpdateSubresource1(ID3D11Resource *pD
 
 				if(tex1)
 				{
-					srcLength = record->Length;
+					srcLength = (size_t)record->Length;
 
 					if(HasDestBox)
 						srcLength = RDCMIN((uint32_t)srcLength, pDstBox->right - pDstBox->left);
